Name mangler: Union getter and setter fields
