+++ hs-bindgen/src-internal/HsBindgen/Backend/Hs/Translation.hs
@@ -33,12 +33,11 @@ import HsBindgen.Backend.Hs.Translation.State qualified as State
 import HsBindgen.Backend.Hs.Translation.Structure
 import HsBindgen.Backend.Hs.Translation.ToFromFunPtr qualified as ToFromFunPtr
 import HsBindgen.Backend.Hs.Translation.Type qualified as Type
+import HsBindgen.Backend.Hs.Translation.Union
 import HsBindgen.Backend.SHs.AST qualified as SHs
 import HsBindgen.Backend.SHs.Translation qualified as SHs
 import HsBindgen.Backend.UniqueSymbol
-import HsBindgen.Config.FixCandidate qualified as FixCandidate
 import HsBindgen.Config.Internal
-import HsBindgen.Errors
 import HsBindgen.Frontend.Analysis.DeclIndex (DeclIndex)
 import HsBindgen.Frontend.Analysis.DeclIndex qualified as DeclIndex
 import HsBindgen.Frontend.AST.Decl qualified as C
@@ -178,7 +177,7 @@ generateDecs uniqueId fns haddockConfig moduleName sizeofs (C.Decl info kind spe
       C.DeclStruct struct -> withCategoryM CType $
         State.immediateM $ structDecs supInsts.struct haddockConfig info struct spec
       C.DeclUnion union -> withCategoryM CType $
-        State.immediateM $ unionDecs fns haddockConfig info union spec
+        State.immediateM $ unionDecs haddockConfig info union spec
       C.DeclEnum enum -> withCategoryM CType $
         State.immediateM $ enumDecs supInsts.enum fns haddockConfig info enum spec
       C.DeclAnonEnumConstant anonEnumConst -> withCategoryM CType $
@@ -253,262 +252,6 @@ opaqueDecs haddockConfig info spec = do
             , kind = Origin.Opaque info.id.cName.name.kind
             , spec = spec
             }
-        }
-
-{-------------------------------------------------------------------------------
-  Unions
--------------------------------------------------------------------------------}
-
-unionDecs ::
-     FieldNamingStrategy
-  -> HaddockConfig
-  -> C.DeclInfo Final
-  -> C.Union Final
-  -> PrescriptiveDeclSpec
-  -> HsM [Hs.Decl]
-unionDecs fieldNaming haddockConfig info union spec = do
-    nt <- newtypeDec
-    flip aux nt <$> State.get
-  where
-    newtypeDec :: HsM Hs.Newtype
-    newtypeDec =
-        Hs.newtypeDec newtypeName newtypeConstr newtypeField
-          newtypeOrigin newtypeComment candidateInsts knownInsts
-      where
-        newtypeName :: Hs.Name Hs.NsTypeConstr
-        newtypeName = Hs.unsafeHsIdHsName info.id.unsafeHsName
-
-        newtypeConstr :: Hs.Name Hs.NsConstr
-        newtypeConstr = union.names.constr
-
-        newtypeField :: Hs.Field
-        newtypeField = Hs.Field {
-              name    = union.names.field
-            , typ     = Hs.HsByteArray
-            , origin  = Origin.GeneratedField
-            , comment = Nothing
-            }
-
-        newtypeOrigin :: Origin.Decl Origin.Newtype
-        newtypeOrigin =  Origin.Decl {
-              info = info
-            , kind = Origin.Union union
-            , spec = spec
-            }
-
-        newtypeComment :: Maybe HsDoc.Comment
-        newtypeComment = mkHaddocks haddockConfig info newtypeName
-
-        candidateInsts :: Set Inst.TypeClass
-        candidateInsts = Set.empty
-
-        knownInsts :: Set Inst.TypeClass
-        knownInsts = Set.fromList $ catMaybes [
-            Just Inst.Generic
-          -- TODO <https://github.com/well-typed/hs-bindgen/issues/1253>
-          -- Should correctly detect 'Inst.HasCBitfield' and 'Inst.HasCField'
-          -- when bit-fields in unions are supported.
-          , if null union.fields then Nothing else Just Inst.HasCField
-          , if null union.fields then Nothing else Just Inst.HasField
-          , Just Inst.ReadRaw
-          , Just Inst.StaticSize
-          , Just Inst.Storable
-          , Just Inst.WriteRaw
-          ]
-
-    -- everything in aux is state-dependent
-    aux :: TranslationState -> Hs.Newtype -> [Hs.Decl]
-    aux transState nt =
-        Hs.DeclNewtype nt : marshalDecls ++ accessorDecls ++
-        concatMap (unionFieldDecls nt.name) union.fields
-      where
-        marshalDecls :: [Hs.Decl]
-        marshalDecls = [
-            Hs.DeclDeriveInstance Hs.DeriveInstance{
-                strategy = Hs.DeriveStock
-              , clss     = Inst.Generic
-              , name     = nt.name
-              , comment  = Nothing
-              }
-          , Hs.DeclDeriveInstance Hs.DeriveInstance{
-                strategy = Hs.DeriveVia sba
-              , clss     = Inst.StaticSize
-              , name     = nt.name
-              , comment  = Nothing
-              }
-          , Hs.DeclDeriveInstance Hs.DeriveInstance{
-                strategy = Hs.DeriveVia sba
-              , clss     = Inst.ReadRaw
-              , name     = nt.name
-              , comment  = Nothing
-              }
-          , Hs.DeclDeriveInstance Hs.DeriveInstance{
-                strategy = Hs.DeriveVia sba
-              , clss     = Inst.WriteRaw
-              , name     = nt.name
-              , comment  = Nothing
-              }
-          , Hs.DeclDeriveInstance Hs.DeriveInstance{
-                strategy = Hs.DeriveVia (HsEquivStorable (Hs.HsTypRef nt.name Nothing))
-              , clss     = Inst.Storable
-              , name     = nt.name
-              , comment  = Nothing
-              }
-          ]
-
-        sba :: Hs.HsType
-        sba =
-            HsSizedByteArray
-              (fromIntegral union.sizeof)
-              (fromIntegral union.alignment)
-
-        accessorDecls :: [Hs.Decl]
-        accessorDecls = concatMap getAccessorDecls union.fields
-
-        getAccessorDecls :: C.UnionField Final -> [Hs.Decl]
-        getAccessorDecls field =
-            if Inst.Storable `Set.notMember` fInsts
-              then []
-              else [
-                  Hs.DeclUnionGetter Hs.UnionGetter{
-                      name    = getterName
-                    , typ     = hsType
-                    , constr  = nt.name
-                    , comment = mkHaddocksFieldInfo haddockConfig info field.info
-                             <> commentRefName (Hs.getName setterName)
-                    }
-                , Hs.DeclUnionSetter Hs.UnionSetter{
-                      name    = setterName
-                    , typ     = hsType
-                    , constr  = nt.name
-                    , comment = commentRefName (Hs.getName getterName)
-                    }
-                ]
-          where
-            hsType     = Type.topLevel field.typ
-            fInsts     = Hs.getInstances
-                            transState.instanceMap
-                            (Just nt.name)
-                            (Set.singleton Inst.Storable)
-                            [hsType]
-
-            -- TODO <https://github.com/well-typed/hs-bindgen/issues/1504>
-            -- This should happen in the name mangler, so that we can deal with
-            -- collisions, errors thrown by 'fixCandidate', etc.
-            --
-            -- With PrefixedFieldNames: field.info.name.hsName already contains the
-            -- type prefix (e.g., "dimPayload_dim2"), so we use it directly.
-            -- With EnableRecordDot: field.info.name.hsName is just the C field name
-            -- (e.g., "dim2"), so we need to add the type name for uniqueness.
-            getterName = Hs.unsafeHsIdHsName $ "get_" <> fieldNameWithPrefix
-            setterName = Hs.unsafeHsIdHsName $ "set_" <> fieldNameWithPrefix
-
-            -- We ensure that we generate the /same/ getter and setter name
-            -- independent of whether record dot syntax is enabled or not.
-            fieldNameWithPrefix :: Hs.Identifier
-            fieldNameWithPrefix = case fieldNaming of
-              PrefixedFieldNames -> field.info.name.hsName
-              EnableRecordDot    ->
-                let candidate :: Text
-                    candidate = Hs.getName nt.name <> "_" <> field.info.name.hsName.text
-
-                    exportedName :: Hs.ExportedName Hs.NsVar
-                    exportedName =
-                       fromMaybe (panicPure $ "Could not construct name for " ++ show candidate) $
-                         FixCandidate.fixCandidate
-                           FixCandidate.fixCandidateDefault
-                           candidate
-
-                in Hs.Identifier exportedName.text
-
-            commentRefName :: Text -> Maybe HsDoc.Comment
-            commentRefName name = Just $ HsDoc.paragraph [
-                HsDoc.Bold [HsDoc.TextContent "See:"]
-              , HsDoc.Identifier name
-              ]
-
--- | 'HasCField' and 'HasField' instances for a field of a
--- union declaration
---
--- Given a union:
---
--- > union myUnion { int option1; char option2 };
---
--- We generate roughly this newtype:
---
--- > newtype MyUnion = MyUnion { unwrapMyUnion :: ByteArray }
---
--- Then, 'unionFieldDecls' will generate roughly the following class instances
--- for the fields @option1@ and @option@ respectively:
---
--- > instance HasCField "myUnion_option1" MyUnion where
--- >   type CFieldType "myUnion_option1" MyUnion = CInt
--- > instance HasField "myUnion_option1" (Ptr MyUnion) (Ptr CInt)
---
--- > instance HasCField "myUnion_option2" MyUnion where
--- >   type CFieldType "myUnion_option2" MyUnion = CChar
--- > instance HasField "myUnion_option2" (Ptr MyUnion) (Ptr CChar)
---
--- This works similarly for bit-fields, but those get a 'HasCBitfield' instance
--- instead of a 'HasCField' instance.
-unionFieldDecls :: Hs.Name Hs.NsTypeConstr -> C.UnionField Final -> [Hs.Decl]
-unionFieldDecls unionName field = [
-      Hs.DeclDefineInstance $
-        Hs.DefineInstance {
-            comment      = Nothing
-          , instanceDecl =
-              case unionFieldWidth field of
-                Nothing -> Hs.InstanceHasCField $ hasCFieldDecl
-                Just w  -> Hs.InstanceHasCBitfield $ hasCBitfieldDecl w
-          }
-    , Hs.DeclDefineInstance $
-        Hs.DefineInstance {
-            comment      = Nothing
-          , instanceDecl = Hs.InstanceHasField hasFieldDecl
-          }
-    ]
-  where
-    -- TODO <https://github.com/well-typed/hs-bindgen/issues/1253>
-    -- Should be changed to @C.unionFieldWidth f@ when bit-fields in unions are
-    -- supported.
-    unionFieldWidth :: C.UnionField Final -> Maybe Int
-    unionFieldWidth _f = Nothing
-
-    parentType :: HsType
-    parentType = Hs.HsTypRef unionName Nothing
-
-    fieldName :: Hs.Name Hs.NsVar
-    fieldName = Hs.unsafeHsIdHsName field.info.name.hsName
-
-    fieldType :: HsType
-    fieldType = Type.topLevel field.typ
-
-    hasFieldDecl :: Hs.HasFieldInstance
-    hasFieldDecl = Hs.HasFieldInstance {
-          parentType = parentType
-        , fieldName  = fieldName
-        , fieldType  = fieldType
-        , deriveVia  =
-            case unionFieldWidth field of
-              Nothing -> Hs.ViaHasCField
-              Just _  -> Hs.ViaHasCBitfield
-        }
-
-    hasCFieldDecl :: Hs.HasCFieldInstance
-    hasCFieldDecl = Hs.HasCFieldInstance {
-          parentType  = parentType
-        , fieldName   = fieldName
-        , cFieldType  = fieldType
-        , fieldOffset = 0
-        }
-
-    hasCBitfieldDecl :: Int -> Hs.HasCBitfieldInstance
-    hasCBitfieldDecl w = Hs.HasCBitfieldInstance {
-          parentType    = parentType
-        , fieldName     = fieldName
-        , cBitfieldType = fieldType
-        , bitOffset     = 0
-        , bitWidth      = w
         }
 
 {-------------------------------------------------------------------------------
